Move dashboard page to its own subdirectory within src/pages and create story for it
